+++ internal/store/badger/badger.go
@@ -1,4 +1,4 @@
-package store
+package badger
 
 import (
 	"encoding/json"
