+++ internal/store/dgraph/dgraph.go
@@ -1,4 +1,4 @@
-package store
+package dgraph
 
 import (
 	"context"
