+++ internal/store/postgres/postgres.go
@@ -1,4 +1,4 @@
-package store
+package postgres
 
 import (
 	"context"
