+++ internal/store/ram/ram.go
@@ -1,4 +1,4 @@
-package store
+package ram
 
 import (
 	"fmt"
