+++ internal/store/store_test.go
@@ -8,19 +8,6 @@ import (
 	"github.com/batazor/shortlink/pkg/link"
 )
 
-var (
-	addLink = link.Link{
-		Url:      "https://example.com",
-		Hash:     "",
-		Describe: "example link",
-	}
-	getLink = link.Link{
-		Url:      "https://example.com",
-		Hash:     "5888cab",
-		Describe: "example link",
-	}
-)
-
 // TestLink ...
 func TestLink(t *testing.T) { //nolint unused
 	ctx := context.Background()
